update customer entity

+++ src/main/java/org/example/springdeployments/repo/CustomerRepo.java
@@ -3,5 +3,5 @@ package org.example.springdeployments.repo;
 import org.example.springdeployments.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepo extends JpaRepository<Customer, Long> {
+public interface CustomerRepo extends JpaRepository<Customer, Integer> {
 }
